extract module
